requirements.txt
# Web 框架
fastapi==0.115.8
uvicorn==0.25.0
httpx==0.27.2
pydantic==2.10.4
python-multipart==0.0.9

# 前端
streamlit==1.38.0

# LangChain & LangGraph
langchain==0.3.15
langchain-openai==0.2.11
langchain-community==0.3.13
langchain-core==0.3.33
langgraph==0.2.69

# 工具
python-dotenv==1.0.1
requests==2.32.2
loguru==0.7.3
tenacity==8.5.0
pandas==2.2.3
numpy==1.26.4

# 股票数据（100%可安装）
tushare==1.4.29